feat: Implement NewsRepository
This commit introduces the `NewsRepository` interface and its implementation, `NewsRepositoryImpl`.

The `NewsRepository` defines a suspend function `getTopHeadlinesTech` to fetch top technology headlines.

The `NewsRepositoryImpl` class implements this interface. It uses the `NewsApi` service to fetch data from the News API, maps the response to the domain model `Article`, and returns a list of articles. It retrieves the `NEWS_API_KEY` from `BuildConfig`.